+++ src/config.ts
@@ -1,0 +1,6 @@
+export const quotes = {
+  none: '',
+  'single-quotes': "'",
+  'double-quotes': '"',
+  backtick: '`'
+} as const;