+++ src/index.ts
@@ -1,3 +1,5 @@
+import { quotes } from './config';
+import { getError, getField, getPropByString } from './helpers';
 import {
 	ArrayType,
 	ConstraintType,
@@ -7,9 +9,12 @@ import {
 	SpecificNumberType,
 	SpecificStringType,
 	Validation,
+	ValidatorConfig
 } from './types';
 
-function validate(rules: Rules, data: Data) {
+const defaultValidatorConfig = {quotes: 'none'} as const
+
+function validate(rules: Rules, data: Data, config: ValidatorConfig = defaultValidatorConfig) {
 	try {
 		let allErrors: string[] | undefined = [];
 		for (let [key, value] of Object.entries(rules)) {
@@ -26,11 +31,15 @@ function validate(rules: Rules, data: Data) {
 			if (key !== '$atleast') {
 				validateSingleData(key, dataToSend, validations, errors);
 			} else {
-				validateAtleastData(data, validations as string[], errors);
+				validateAtleastData(data, validations as Validation[], errors);
 			}
 
 			allErrors.push(...errors);
 		}
+
+		const quote = quotes[config.quotes ?? 'none'];
+		allErrors = allErrors.map(e => e.replace(/"/g, quote))
+		allErrors = [...new Set(allErrors)];
 
 		if (allErrors.length === 0) {
 			allErrors = undefined;
@@ -62,7 +71,7 @@ function validateSingleData(key: string, value: any, validations: Validation[], 
 
 		// ! string,number,boolean,object,array,bigint
 		if ('string,number,boolean,object,array,bigint'.split(',').includes(validation)) {
-			checkDataType(key, value, validation as DataType, previousValidations, errors);
+			checkDataType(key, value, validation as DataType, previousValidations, validations, errors);
 			if (errors.length) {
 				break;
 			}
@@ -74,7 +83,7 @@ function validateSingleData(key: string, value: any, validations: Validation[], 
 				.split(',')
 				.includes(validation)
 		) {
-			checkSpecificStringType(key, value, validation as SpecificStringType, previousValidations, errors);
+			checkSpecificStringType(key, value, validation as SpecificStringType, previousValidations, validations, errors);
 			if (errors.length) {
 				break;
 			}
@@ -82,7 +91,7 @@ function validateSingleData(key: string, value: any, validations: Validation[], 
 
 		// ! int,positive,negative,natural,whole
 		if ('int,positive,negative,natural,whole'.split(',').includes(validation)) {
-			checkSpecificNumberType(key, value, validation as SpecificNumberType, previousValidations, errors);
+			checkSpecificNumberType(key, value, validation as SpecificNumberType, previousValidations, validations, errors);
 			if (errors.length) {
 				break;
 			}
@@ -94,7 +103,7 @@ function validateSingleData(key: string, value: any, validations: Validation[], 
 				.split(',')
 				.some((e) => validation.startsWith(e + ':'))
 		) {
-			checkConstraint(key, value, validation as ConstraintType, previousValidations, errors);
+			checkConstraint(key, value, validation as ConstraintType, previousValidations, validations, errors);
 			if (errors.length) {
 				break;
 			}
@@ -107,6 +116,7 @@ function validateSingleData(key: string, value: any, validations: Validation[], 
 				value,
 				validation as ArrayType,
 				previousValidations,
+				validations,
 				errors,
 				optionalArrays,
 				nullableArrays
@@ -118,14 +128,15 @@ function validateSingleData(key: string, value: any, validations: Validation[], 
 	}
 }
 
-function validateAtleastData(data: Data, keys: string[], errors: string[]) {
-	for (let key of keys) {
+function validateAtleastData(data: Data, validations: Validation[], errors: string[]) {
+	let error = getError(validations)
+	for (let key of validations) {
 		let value = !key.includes('.') ? data[key] : getPropByString(data, key);
 		if (value !== null && value !== undefined) {
 			return;
 		}
 	}
-	errors.push(`at least one of ${keys.map((e) => `"${e}"`).join(', ')} is required`);
+	errors.push(error ?? `at least one of ${validations.map((e) => `"${e}"`).join(', ')} is required`);
 }
 
 function checkDataType(
@@ -133,53 +144,57 @@ function checkDataType(
 	value: any,
 	dataType: DataType,
 	previousValidations: Validation[],
+	validations: Validation[],
 	errors: string[]
 ) {
+	let field = getField(validations, key);
+	let error = getError(validations);
+
 	if (value === undefined || value === null) {
-		errors.push(`"${key}" is required`);
+		errors.push(error ?? `"${field ?? key}" is required`);
 		return;
 	}
 
 	if (dataType === 'symbol' && typeof value !== 'symbol') {
-		errors.push(`"${key}" must be symbol`);
+		errors.push(error ?? `"${field ?? key}" must be symbol`);
 		return;
 	}
 
 	const hasString = previousValidations.includes('string') || previousValidations.includes('arrayof:string')
 
 	if (dataType === 'string' && typeof value !== 'string') {
-		errors.push(`"${key}" must be string`);
+		errors.push(error ?? `"${field ?? key}" must be string`);
 		return;
 	}
 
 	if (dataType === 'number' && hasString && Number.isNaN(+value)) {
-		errors.push(`"${key}" must be a valid numeric string`);
+		errors.push(error ?? `"${field ?? key}" must be a valid numeric string`);
 		return;
 	} else if (dataType === 'number' && !hasString && typeof value !== 'number') {
-		errors.push(`"${key}" must be a valid number`);
+		errors.push(error ?? `"${field ?? key}" must be a valid number`);
 		return;
 	}
 
 	if (dataType === 'bigint' && typeof value !== 'bigint') {
-		errors.push(`"${key}" must be bigint`);
+		errors.push(error ?? `"${field ?? key}" must be bigint`);
 		return;
 	}
 
 	if (dataType === 'boolean' && hasString && !['true', 'false'].includes(value)) {
-		errors.push(`"${key}" must be a valid boolean string`);
+		errors.push(error ?? `"${field ?? key}" must be a valid boolean string`);
 		return;
 	} else if (dataType === 'boolean' && !hasString && typeof value !== 'boolean') {
-		errors.push(`"${key}" must be a valid boolean`);
+		errors.push(error ?? `"${field ?? key}" must be a valid boolean`);
 		return;
 	}
 
 	if (dataType === 'array' && !Array.isArray(value)) {
-		errors.push(`"${key}" must be an array`);
+		errors.push(error ?? `"${field ?? key}" must be an array`);
 		return;
 	}
 
 	if (dataType === 'object' && (Array.isArray(value) || typeof value !== 'object')) {
-		errors.push(`"${key}" must be an object`);
+		errors.push(error ?? `"${field ?? key}" must be an object`);
 		return;
 	}
 }
@@ -189,14 +204,18 @@ function checkSpecificStringType(
 	value: any,
 	specificType: SpecificStringType,
 	previousValidations: Validation[],
+	validations: Validation[],
 	errors: string[]
 ) {
+	let field = getField(validations, key);
+	let error = getError(validations);
+
 	if (value === undefined || value === null) {
-		errors.push(`"${key}" is required`);
+		errors.push(error ?? `"${field ?? key}" is required`);
 		return;
 	}
 
-	checkDataType(key, value, 'string', previousValidations, errors);
+	checkDataType(key, value, 'string', previousValidations, validations, errors);
 	if (errors.length) {
 		return;
 	}
@@ -205,7 +224,7 @@ function checkSpecificStringType(
 		specificType === 'email' &&
 		!/^[A-Z0-9_'%=+!`#~$*?^{}&|-]+([\.][A-Z0-9_'%=+!`#~$*?^{}&|-]+)*@[A-Z0-9-]+(\.[A-Z0-9-]+)+$/i.test(value)
 	) {
-		errors.push(`"${key}" must be a valid email`);
+		errors.push(error ?? `"${field ?? key}" must be a valid email`);
 		return;
 	}
 
@@ -215,7 +234,7 @@ function checkSpecificStringType(
 			value
 		)
 	) {
-		errors.push(`"${key}" must be a valid url`);
+		errors.push(error ?? `"${field ?? key}" must be a valid url`);
 		return;
 	}
 
@@ -223,42 +242,42 @@ function checkSpecificStringType(
 		specificType === 'domain' &&
 		!/^[a-zA-Z0-9][a-zA-Z0-9-_]{0,61}[a-zA-Z0-9]{0,1}\.([a-zA-Z]{1,6}|[a-zA-Z0-9-]{1,30}\.[a-zA-Z]{2,3})$/.test(value)
 	) {
-		errors.push(`"${key}" must be a valid domain`);
+		errors.push(error ?? `"${field ?? key}" must be a valid domain`);
 		return;
 	}
 
-	if (specificType === 'name' && !/^[A-Za-z. ]{2,30}$/.test(value)) {
-		errors.push(`"${key}" must be a valid name`);
+	if (specificType === 'name' && !/^([a-z]+[,.]?[ ]?|[a-z]+['-]?)+$/.test(value)) {
+		errors.push(error ?? `"${field ?? key}" must be a valid name`);
 		return;
 	}
 
 	if (specificType === 'fullname' && !/^([a-zA-Z]{2,}\s[a-zA-Z]{1,}'?-?[a-zA-Z]{2,}\s?([a-zA-Z]{1,})?)$/.test(value)) {
-		errors.push(`"${key}" must be a valid fullname`);
+		errors.push(error ?? `"${field ?? key}" must be a valid fullname`);
 		return;
 	}
 
 	if (specificType === 'username' && !/^[^\W_](?!.*?[._]{2})[\w.]{6,18}[^\W_]$/.test(value)) {
-		errors.push(`"${key}" must be a valid username`);
+		errors.push(error ?? `"${field ?? key}" must be a valid username`);
 		return;
 	}
 
 	if (specificType === 'alpha' && !/^[A-Za-z]{1,}$/.test(value)) {
-		errors.push(`"${key}" must be a valid alpha`);
+		errors.push(error ?? `"${field ?? key}" must be a valid alpha`);
 		return;
 	}
 
 	if (specificType === 'alphanumeric' && !/^[A-Za-z0-9]{1,}$/.test(value)) {
-		errors.push(`"${key}" must be a valid alphanumeric`);
+		errors.push(error ?? `"${field ?? key}" must be a valid alphanumeric`);
 		return;
 	}
 
 	if (specificType === 'phone' && !/^(?:\+\d{1,3}\s?)?(?:\(\d+\))?(?:\d+\s?)+(?:\d{1,4})$/.test(value)) {
-		errors.push(`"${key}" must be a valid phone`);
+		errors.push(error ?? `"${field ?? key}" must be a valid phone`);
 		return;
 	}
 
 	if (specificType === 'mongoid' && !/^[0-9a-fA-F]{24}$/.test(value)) {
-		errors.push(`"${key}" must be a valid mongodb id`);
+		errors.push(error ?? `"${field ?? key}" must be a valid mongodb id`);
 		return;
 	}
 
@@ -268,27 +287,27 @@ function checkSpecificStringType(
 			value
 		)
 	) {
-		errors.push(`"${key}" must be a valid date`);
+		errors.push(error ?? `"${field ?? key}" must be a valid date`);
 		return;
 	}
 
 	if (specificType === 'dateonly' && !/^(\d{4})-(0[1-9]|1[0-2])-([12]\d|0[1-9]|3[01])$/.test(value)) {
-		errors.push(`"${key}" must be a valid date`);
+		errors.push(error ?? `"${field ?? key}" must be a valid date`);
 		return;
 	}
 
 	if (specificType === 'time' && !/^(?:[01]\d|2[0-3]):([0-5]\d)(?::([0-5]\d))?(?:\.\d{1,4})?$/.test(value)) {
-		errors.push(`"${key}" must be a valid time`);
+		errors.push(error ?? `"${field ?? key}" must be a valid time`);
 		return;
 	}
 
 	if (specificType === 'lower' && !/^[^A-Z]+$/.test(value)) {
-		errors.push(`"${key}" must not contains upper case letters`);
+		errors.push(error ?? `"${field ?? key}" must not contains upper case letters`);
 		return;
 	}
 
 	if (specificType === 'upper' && !/^[^a-z]+$/.test(value)) {
-		errors.push(`"${key}" must not contains lower case letters`);
+		errors.push(error ?? `"${field ?? key}" must not contains lower case letters`);
 		return;
 	}
 
@@ -296,7 +315,7 @@ function checkSpecificStringType(
 		specificType === 'ip' &&
 		!/^(?:(?:25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)\.){3}(?:25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)$/.test(value)
 	) {
-		errors.push(`"${key}" must be a valid IP address`);
+		errors.push(error ?? `"${field ?? key}" must be a valid IP address`);
 		return;
 	}
 }
@@ -306,14 +325,18 @@ function checkSpecificNumberType(
 	value: any,
 	specificType: SpecificNumberType,
 	previousValidations: Validation[],
+	validations: Validation[],
 	errors: string[]
 ) {
+	let field = getField(validations, key);
+	let error = getError(validations);
+
 	if (value === undefined || value === null) {
-		errors.push(`"${key}" is required`);
+		errors.push(error ??`"${field ?? key}" is required`);
 		return;
 	}
 
-	checkDataType(key, value, 'number', previousValidations, errors);
+	checkDataType(key, value, 'number', previousValidations, validations, errors);
 	if (errors.length) {
 		return;
 	}
@@ -321,42 +344,42 @@ function checkSpecificNumberType(
 	const hasString = previousValidations.includes('string') || previousValidations.includes('arrayof:string')
 
 	if (specificType === 'int' && hasString && `${+value}`.includes('.')) {
-		errors.push(`"${key}" must be a valid integer string`);
+		errors.push(error ??`"${field ?? key}" must be a valid integer string`);
 		return;
 	} else if (specificType === 'int' && !hasString && `${value}`.includes('.')) {
-		errors.push(`"${key}" must be a valid integer`);
+		errors.push(error ??`"${field ?? key}" must be a valid integer`);
 		return;
 	}
 
 	if (specificType === 'positive' && hasString && +value <= 0) {
-		errors.push(`"${key}" must be a valid positive numeric string`);
+		errors.push(error ??`"${field ?? key}" must be a valid positive numeric string`);
 		return;
 	} else if (specificType === 'positive' && !hasString && value <= 0) {
-		errors.push(`"${key}" must be a valid positive number`);
+		errors.push(error ??`"${field ?? key}" must be a valid positive number`);
 		return;
 	}
 
 	if (specificType === 'negative' && hasString && +value >= 0) {
-		errors.push(`"${key}" must be a valid negative numeric string`);
+		errors.push(error ??`"${field ?? key}" must be a valid negative numeric string`);
 		return;
 	} else if (specificType === 'negative' && !hasString && value >= 0) {
-		errors.push(`"${key}" must be a valid negative number`);
+		errors.push(error ??`"${field ?? key}" must be a valid negative number`);
 		return;
 	}
 
 	if (specificType === 'natural' && hasString && (`${+value}`.includes('.') || +value <= 0)) {
-		errors.push(`"${key}" must be a valid natural numeric string`);
+		errors.push(error ??`"${field ?? key}" must be a valid natural numeric string`);
 		return;
 	} else if (specificType === 'natural' && !hasString && (`${value}`.includes('.') || value <= 0)) {
-		errors.push(`"${key}" must be a valid natural number`);
+		errors.push(error ??`"${field ?? key}" must be a valid natural number`);
 		return;
 	}
 
 	if (specificType === 'whole' && hasString && (`${+value}`.includes('.') || +value < 0)) {
-		errors.push(`"${key}" must be a valid whole numeric string`);
+		errors.push(error ??`"${field ?? key}" must be a valid whole numeric string`);
 		return;
 	} else if (specificType === 'whole' && !hasString && (`${value}`.includes('.') || value < 0)) {
-		errors.push(`"${key}" must be a valid whole number`);
+		errors.push(error ??`"${field ?? key}" must be a valid whole number`);
 		return;
 	}
 }
@@ -366,10 +389,14 @@ function checkConstraint(
 	value: any,
 	type: ConstraintType,
 	previousValidations: Validation[],
+	validations: Validation[],
 	errors: string[]
 ) {
+	let field = getField(validations, key);
+	let error = getError(validations);
+	
 	if (value === undefined || value === null) {
-		errors.push(`"${key}" is required`);
+		errors.push(error ?? `"${field ?? key}" is required`);
 		return;
 	}
 
@@ -382,18 +409,18 @@ function checkConstraint(
 		let data = type.substring(6);
 
 		if (isString && data !== value) {
-			errors.push(`"${key}" must be equal to ${data}`);
+			errors.push(error ?? `"${field ?? key}" must be equal to ${data}`);
 			return;
 		}
 
 		if (isNumber && +data !== value) {
-			errors.push(`"${key}" must be equal to ${data}`);
+			errors.push(error ?? `"${field ?? key}" must be equal to ${data}`);
 			return;
 		}
 
 		if (typeof value === 'boolean') {
 			if ((data === 'true' && value === false) || (data === 'false' && value === true))
-				errors.push(`"${key}" must be equal to ${data}`);
+				errors.push(error ?? `"${field ?? key}" must be equal to ${data}`);
 			return;
 		}
 	}
@@ -403,12 +430,12 @@ function checkConstraint(
 		let size = +type.substring(5);
 
 		if (isString && !isNumeric && value.length !== size) {
-			errors.push(`"${key}" must have length ${size}`);
+			errors.push(error ?? `"${field ?? key}" must have length ${size}`);
 			return;
 		}
 
 		if (Array.isArray(value) && value.length !== size) {
-			errors.push(`"${key}" must have length ${size}`);
+			errors.push(error ?? `"${field ?? key}" must have length ${size}`);
 			return;
 		}
 
@@ -420,7 +447,7 @@ function checkConstraint(
 				.filter((e) => ['e', '-', '+', '.'].includes(e))
 				.join('').length;
 			if (valueLength !== size) {
-				errors.push(`"${key}" must have ${size} digits`);
+				errors.push(error ?? `"${field ?? key}" must have ${size} digits`);
 				return;
 			}
 		}
@@ -436,22 +463,22 @@ function checkConstraint(
 		if (min instanceof Date) {
 			let date = new Date(Date.parse(value));
 			if (date < min) {
-				errors.push(`"${key}" must be at least ${min.toISOString()}`);
+				errors.push(error ?? `"${field ?? key}" must be at least ${min.toISOString()}`);
 				return;
 			}
 		} else {
 			if (isString && !isNumeric && value.length < min) {
-				errors.push(`"${key}" must have length of at least ${min}`);
+				errors.push(error ?? `"${field ?? key}" must have length of at least ${min}`);
 				return;
 			}
 
 			if (Array.isArray(value) && value.length < min) {
-				errors.push(`"${key}" must have length of at least ${min}`);
+				errors.push(error ?? `"${field ?? key}" must have length of at least ${min}`);
 				return;
 			}
 
 			if ( ((isString && isNumeric) || isNumber) && +value < min) {
-				errors.push(`"${key}" must be at least ${min}`);
+				errors.push(error ?? `"${field ?? key}" must be at least ${min}`);
 				return;
 			}
 		}
@@ -467,22 +494,22 @@ function checkConstraint(
 		if (max instanceof Date) {
 			let date = new Date(Date.parse(value));
 			if (date > max) {
-				errors.push(`"${key}" must be at most ${max.toISOString()}`);
+				errors.push(error ?? `"${field ?? key}" must be at most ${max.toISOString()}`);
 				return;
 			}
 		} else {
 			if (isString && !isNumeric &&  value.length > max) {
-				errors.push(`"${key}" must have length of at most ${max}`);
+				errors.push(error ?? `"${field ?? key}" must have length of at most ${max}`);
 				return;
 			}
 
 			if (Array.isArray(value) && value.length > max) {
-				errors.push(`"${key}" must have length of at most ${max}`);
+				errors.push(error ?? `"${field ?? key}" must have length of at most ${max}`);
 				return;
 			}
 
 			if (((isString && isNumeric) || isNumber) && +value > max) {
-				errors.push(`"${key}" must be at most ${max}`);
+				errors.push(error ?? `"${field ?? key}" must be at most ${max}`);
 				return;
 			}
 		}
@@ -497,12 +524,12 @@ function checkConstraint(
 		let regex = new RegExp(main, flags);
 
 		if (typeof value !== 'string') {
-			errors.push(`"${key}" must be of type string`);
+			errors.push(error ?? `"${field ?? key}" must be of type string`);
 			return;
 		}
 
 		if (!regex.test(value)) {
-			errors.push(`"${key}" is invalid`);
+			errors.push(error ?? `"${field ?? key}" is invalid`);
 			return;
 		}
 	}
@@ -514,11 +541,11 @@ function checkConstraint(
 		if (typeof value === 'string') {
 			let n = +value;
 			if (Number.isNaN(n)) {
-				errors.push(`"${key}" must be a valid numeric string`);
+				errors.push(error ?? `"${field ?? key}" must be a valid numeric string`);
 				return;
 			}
 			if (!value.includes('.') && size > 0) {
-				errors.push(`"${key}" must have ${size} decimal places`);
+				errors.push(error ?? `"${field ?? key}" must have ${size} decimal places`);
 				return;
 			}
 			let index = value.lastIndexOf('.');
@@ -529,19 +556,19 @@ function checkConstraint(
 			}
 
 			if (digits.length !== size) {
-				errors.push(`"${key}" must have ${size} decimal places`);
+				errors.push(error ?? `"${field ?? key}" must have ${size} decimal places`);
 				return;
 			}
 		}
 
 		if (typeof value === 'number') {
 			if (Number.isNaN(value)) {
-				errors.push(`"${key}" must be a value number`);
+				errors.push(error ?? `"${field ?? key}" must be a value number`);
 				return;
 			}
 			let str = `${value}`;
 			if (!str.includes('.') && size > 0) {
-				errors.push(`"${key}" must have ${size} decimal places`);
+				errors.push(error ?? `"${field ?? key}" must have ${size} decimal places`);
 				return;
 			}
 			if (!str.includes('.') && size === 0) {
@@ -551,7 +578,7 @@ function checkConstraint(
 			let index = str.lastIndexOf('.');
 			let digits = str.substring(index + 1);
 			if (digits.length !== size) {
-				errors.push(`"${key}" must have ${size} decimal places`);
+				errors.push(error ?? `"${field ?? key}" must have ${size} decimal places`);
 				return;
 			}
 		}
@@ -564,11 +591,11 @@ function checkConstraint(
 		if (typeof value === 'string') {
 			let n = +value;
 			if (Number.isNaN(n)) {
-				errors.push(`"${key}" must be a valid numeric string`);
+				errors.push(error ?? `"${field ?? key}" must be a valid numeric string`);
 				return;
 			}
 			if (!value.includes('.') && min > 0) {
-				errors.push(`"${key}" must have at least ${min} decimal places`);
+				errors.push(error ?? `"${field ?? key}" must have at least ${min} decimal places`);
 				return;
 			}
 			let index = value.lastIndexOf('.');
@@ -579,19 +606,19 @@ function checkConstraint(
 			}
 
 			if (digits.length < min) {
-				errors.push(`"${key}" must have at least ${min} decimal places`);
+				errors.push(error ?? `"${field ?? key}" must have at least ${min} decimal places`);
 				return;
 			}
 		}
 
 		if (typeof value === 'number') {
 			if (Number.isNaN(value)) {
-				errors.push(`"${key}" must be a value number`);
+				errors.push(error ?? `"${field ?? key}" must be a value number`);
 				return;
 			}
 			let str = `${value}`;
 			if (!str.includes('.') && min > 0) {
-				errors.push(`"${key}" must have at least ${min} decimal places`);
+				errors.push(error ?? `"${field ?? key}" must have at least ${min} decimal places`);
 				return;
 			}
 			if (!str.includes('.') && min === 0) {
@@ -602,7 +629,7 @@ function checkConstraint(
 			let digits = str.substring(index + 1);
 
 			if (digits.length < min) {
-				errors.push(`"${key}" must have at least ${min} decimal places`);
+				errors.push(error ?? `"${field ?? key}" must have at least ${min} decimal places`);
 				return;
 			}
 		}
@@ -615,7 +642,7 @@ function checkConstraint(
 		if (typeof value === 'string') {
 			let n = +value;
 			if (Number.isNaN(n)) {
-				errors.push(`"${key}" must be a valid numeric string`);
+				errors.push(error ?? `"${field ?? key}" must be a valid numeric string`);
 				return;
 			}
 
@@ -627,14 +654,14 @@ function checkConstraint(
 			let digits = value.substring(index + 1);
 
 			if (digits.length > max) {
-				errors.push(`"${key}" must have at most ${max} decimal places`);
+				errors.push(error ?? `"${field ?? key}" must have at most ${max} decimal places`);
 				return;
 			}
 		}
 
 		if (typeof value === 'number') {
 			if (Number.isNaN(value)) {
-				errors.push(`"${key}" must be a value number`);
+				errors.push(error ?? `"${field ?? key}" must be a value number`);
 				return;
 			}
 			let str = `${value}`;
@@ -646,7 +673,7 @@ function checkConstraint(
 			let index = str.lastIndexOf('.');
 			let digits = str.substring(index + 1);
 			if (digits.length > max) {
-				errors.push(`"${key}" must have at most ${max} decimal places`);
+				errors.push(error ?? `"${field ?? key}" must have at most ${max} decimal places`);
 				return;
 			}
 		}
@@ -658,7 +685,7 @@ function checkConstraint(
 
 		if (isString && !isNumeric) {
 			if (!array.includes(value)) {
-				errors.push(`"${key}" is invalid`);
+				errors.push(error ?? `"${field ?? key}" is invalid`);
 				return;
 			}
 		}
@@ -670,7 +697,7 @@ function checkConstraint(
 					.map((e) => +e)
 					.includes(+value)
 			) {
-				errors.push(`"${key}" is invalid`);
+				errors.push(error ?? `"${field ?? key}" is invalid`);
 				return;
 			}
 		}
@@ -682,7 +709,7 @@ function checkConstraint(
 					.map((e) => e === 'true')
 					.includes(value)
 			) {
-				errors.push(`"${key}" is invalid`);
+				errors.push(error ?? `"${field ?? key}" is invalid`);
 				return;
 			}
 		}
@@ -694,14 +721,18 @@ function checkSpecificArrayType(
 	value: any,
 	type: ArrayType,
 	previousValidations: Validation[],
+	validations: Validation[],
 	errors: string[],
 	optionalArrays: any[][],
 	nullableArrays: any[][]
 ) {
+	let field = getField(validations, key);
+	let error = getError(validations);
+
 	let validation = type.substring(8) as Validation;
 
 	if (!Array.isArray(value)) {
-		errors.push(`"${key}" must be an array`);
+		errors.push(error ?? `"${field ?? key}" must be an array`);
 		return;
 	}
 
@@ -721,7 +752,7 @@ function checkSpecificArrayType(
 	for (let index = 0; index <= array.length - 1; index++) {
 		let element = array[index];
 		let newErrors = [] as string[];
-		let elementKey = `${key}[${index}]`;
+		let elementKey = `${field ?? key}[${index}]`;
 
 		if ((isOptional && element === undefined) || (isNullable && element === null)) {
 			continue;
@@ -729,7 +760,7 @@ function checkSpecificArrayType(
 
 		// ! string,number,boolean,array,object,bigint,symbol
 		if ('string,number,boolean,array,object,bigint,symbol'.split(',').includes(validation)) {
-			checkDataType(elementKey, element, validation as DataType, previousValidations, newErrors);
+			checkDataType(elementKey, element, validation as DataType, previousValidations, validations, newErrors);
 		}
 
 		// ! email,url,domain,name,fullname,username,alpha,alphanumeric,phone,mongoid,date,dateonly,time,lower,upper,ip
@@ -738,22 +769,22 @@ function checkSpecificArrayType(
 				.split(',')
 				.includes(validation)
 		) {
-			checkDataType(elementKey, element, 'string', previousValidations, newErrors);
+			checkDataType(elementKey, element, 'string', previousValidations, validations, newErrors);
 			if (newErrors.length) {
 				errors.push(...newErrors);
 				continue;
 			}
-			checkSpecificStringType(elementKey, element, validation as SpecificStringType, previousValidations, newErrors);
+			checkSpecificStringType(elementKey, element, validation as SpecificStringType, previousValidations, validations, newErrors);
 		}
 
 		// ! int,positive,negative,natural,whole
 		if ('int,positive,negative,natural,whole'.split(',').includes(validation)) {
-			checkDataType(elementKey, element, 'number', previousValidations, newErrors);
+			checkDataType(elementKey, element, 'number', previousValidations, validations, newErrors);
 			if (newErrors.length) {
 				errors.push(...newErrors);
 				continue;
 			}
-			checkSpecificNumberType(elementKey, element, validation as SpecificNumberType, previousValidations, newErrors);
+			checkSpecificNumberType(elementKey, element, validation as SpecificNumberType, previousValidations, validations, newErrors);
 		}
 
 		// ! equal,size,min,max,regex,decimalsize,decimalmin,decimalmax,enums
@@ -762,7 +793,7 @@ function checkSpecificArrayType(
 				.split(',')
 				.some((e) => validation.startsWith(`${e}:`))
 		) {
-			checkConstraint(elementKey, element, validation as ConstraintType, previousValidations, newErrors);
+			checkConstraint(elementKey, element, validation as ConstraintType, previousValidations, validations, newErrors);
 		}
 
 		// // ! Nested Array Check
@@ -788,26 +819,6 @@ function checkSpecificArrayType(
 	}
 }
 
-function getPropByString(obj: Data, propString: string) {
-	if (!propString) return obj;
-
-	let prop,
-		props = propString.split('.');
-	let i = 0;
-
-	for (let iLen = props.length - 1; i < iLen; i++) {
-		prop = props[i];
-
-		let candidate = obj[prop];
-		if (candidate !== undefined) {
-			obj = candidate;
-		} else {
-			break;
-		}
-	}
-	return obj[props[i]];
-}
-
 const Validator = { validate };
 
-module.exports = Validator;
+export default Validator;